extracted enttec package info as much as possible
